+++ R/ard_effectsize_hedges_g.R
@@ -65,7 +65,7 @@ ard_effectsize_hedges_g <- function(data, by, variables, conf.level = 0.95, ...)
 
   # return empty ARD if no variables selected ----------------------------------
   if (is_empty(variables)) {
-    return(dplyr::tibble() |> cards::as_card())
+    return(dplyr::tibble() |> cards::as_card(check=FALSE))
   }
 
   # build ARD ------------------------------------------------------------------
@@ -117,7 +117,7 @@ ard_effectsize_paired_hedges_g <- function(data, by, variables, id, conf.level =
 
   # return empty ARD if no variables selected ----------------------------------
   if (is_empty(variables)) {
-    return(dplyr::tibble() |> cards::as_card())
+    return(dplyr::tibble() |> cards::as_card(check=FALSE))
   }
 
   # build ARD ------------------------------------------------------------------
@@ -190,6 +190,6 @@ ard_effectsize_paired_hedges_g <- function(data, by, variables, id, conf.level =
       by = "stat_name"
     ) |>
     dplyr::mutate(stat_label = dplyr::coalesce(.data$stat_label, .data$stat_name)) |>
-    cards::as_card() |>
+    cards::as_card(check=FALSE) |>
     cards::tidy_ard_column_order()
 }