+++ tests/testthat/test-ard_continuous_ci.survey.design.R
@@ -35,7 +35,7 @@ test_that("ard_continuous_ci(variables)", {
 
   expect_equal(
     ard_continuous_ci(dclus1, variables = starts_with("xxxxxx")),
-    dplyr::tibble() |> cards::as_card()
+    dplyr::tibble() |> cards::as_card(check=FALSE)
   )
 
   # check NA values don't affect result