+++ tests/testthat/test-ard_stats_wilcox_test_onesample.R
@@ -47,7 +47,7 @@ test_that("ard_stats_wilcox_test_onesample() works", {
       cards::ADSL,
       variables = character(0)
     ),
-    dplyr::tibble() |> cards::as_card()
+    dplyr::tibble() |> cards::as_card(check=FALSE)
   )
 })