+++ NEWS.md
@@ -1,5 +1,7 @@
 # cardx 0.3.2.9000
 
+* Added fix to ensure `as_card` does not error after update to `cards`
+
 # cardx 0.3.2
 
 * Swapped internal use of `dplyr::case_when()` for `dplyr::recode_values()` as the former is now deprecated. (#327)